+++ commerce-domain/src/main/java/co/hooghly/commerce/repository/LanguageRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.hooghly.commerce.domain.Language;
 
-public interface LanguageRepository extends JpaRepository <Language, Integer> {
+public interface LanguageRepository extends JpaRepository <Language, Long> {
 	
 	Language findByCode(String code);
